Folder structure refactor. Tab toggle name change in event and following pages. Groups and user tabs section grid change.
