+++ src/app/(main)/event/[id]/page.tsx
