+++ src/app/(main)/group/[id]/page.tsx
