+++ src/app/components/groups/GroupsTab.tsx
@@ -11,7 +11,6 @@ const GroupsTab: FC = ({ groups }: object[]) => {
           item
           xs={12}
           sm={6}
-          md={4}
           lg={3}
           className="flex justify-center"
         >
