+++ src/app/components/layout/header/Header.tsx
@@ -24,13 +24,13 @@ const Header: FC = () => {
                 </h1>
               </div>
             </div>
-            <div className="ml-auto mr-3">
+            <div className="ml-auto mr-2">
               <Button variant="plain" className="p-1">
                 <SearchIcon className="fill-primary-rose dark:fill-dark-primary-light-blue h-6 w-6" />
               </Button>
             </div>
             <div className="hidden md:block">
-              <div className="ml-4 flex items-center md:ml-6">
+              <div className="ml-4 flex items-center md:ml-0">
                 <Button variant="plain" className="p-1">
                   <Badge badgeContent={4} size="sm" className="text-xs">
                     <Typography>
