+++ src/app/globals.css
@@ -61,11 +61,15 @@
 
 /* List item button */
 .MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain {
-    @apply bg-primary dark:bg-dark-primary hover:bg-primary-rose dark:hover:bg-dark-primary-blue text-primary-rose dark:text-dark-primary-light-blue hover:text-primary hover:text-dark-primary-light-blue;
+    @apply bg-primary dark:bg-dark-primary hover:bg-secondary dark:hover:bg-dark-primary-blue text-primary-rose dark:text-dark-primary-light-blue hover:text-primary hover:text-dark-primary-light-blue;
 }
 
 .MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain[aria-active="true"] {
-    @apply bg-primary-rose text-primary dark:text-dark-primary-light-blue dark:bg-dark-primary-blue;
+    @apply bg-primary-rose text-primary dark:text-dark-primary-light-blue dark:bg-dark-primary-blue hover:bg-secondary dark:hover:bg-dark-primary-light-blue;
+}
+
+.MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain[aria-active="true"]:hover {
+    @apply dark:text-dark-primary;
 }
 
 .MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain svg {
@@ -74,6 +78,10 @@
 
 .MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain[aria-active="true"] svg {
     @apply fill-primary dark:fill-dark-primary-light-blue;
+}
+
+.MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain[aria-active="true"]:hover svg {
+    @apply dark:fill-dark-primary;
 }
 
 .MuiListItemButton-colorNeutral.MuiListItemButton-variantPlain:hover svg {
